fix: reduce Supabase auth rate limit by narrowing middleware matcher
The middleware was calling supabase.auth.getUser() on every single
request site-wide (via an overly broad matcher), triggering one
network call to Supabase /auth/v1/user per page load. This caused
the Supabase rate limit to be hit under normal usage.

Changes:
- Narrow the matcher in src/middleware.ts to only cover protected
  routes (/dashboard, /labs, /perfil) and auth routes (/login, /register)
- Add proper route-protection redirects (unauthenticated → /login,
  authenticated → /forum for auth pages)
- Remove redundant apps/frontend/middleware.ts (outside src/ directory,
  never executed by Next.js in a src/-based project)
