--- apps/frontend/middleware.ts
@@ -1,59 +1,0 @@
-import { createServerClient } from '@supabase/ssr';
-import { NextResponse, type NextRequest } from 'next/server';
-
-export async function middleware(request: NextRequest) {
-  let supabaseResponse = NextResponse.next({ request });
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        getAll() {
-          return request.cookies.getAll();
-        },
-        setAll(cookiesToSet) {
-          cookiesToSet.forEach(({ name, value }) =>
-            request.cookies.set(name, value)
-          );
-          supabaseResponse = NextResponse.next({ request });
-          cookiesToSet.forEach(({ name, value, options }) =>
-            supabaseResponse.cookies.set(name, value, options)
-          );
-        },
-      },
-    }
-  );
-
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  const { pathname } = request.nextUrl;
-
-  // Redirigir a /login si no hay sesión en rutas protegidas
-  if (!user && (pathname.startsWith('/dashboard') || pathname.startsWith('/labs'))) {
-    const loginUrl = request.nextUrl.clone();
-    loginUrl.pathname = '/login';
-    loginUrl.searchParams.set('redirectedFrom', pathname);
-    return NextResponse.redirect(loginUrl);
-  }
-
-  // Redirigir a /forum si ya está autenticado e intenta ir a /login o /register
-  if (user && (pathname === '/login' || pathname === '/register')) {
-    const forumUrl = request.nextUrl.clone();
-    forumUrl.pathname = '/forum';
-    return NextResponse.redirect(forumUrl);
-  }
-
-  return supabaseResponse;
-}
-
-export const config = {
-  matcher: [
-    '/dashboard/:path*',
-    '/labs/:path*',
-    '/login',
-    '/register',
-  ],
-};
